% Generated by roxygen2 (4.0.1): do not edit by hand
\name{MERRA.get.slab.from.hourly}
\alias{MERRA.get.slab.from.hourly}
\title{Extract a hyperslab of data.}
\usage{
MERRA.get.slab.from.hourly(variable, date.range, height.range = NULL,
  lat.range = c(-90, 90), lon.range = c(-180, 180), opendap = TRUE,
  type = "mean")
}
\arguments{
\item{variable}{MERRA variable name, only 2d variables will work.}

\item{date.range}{A pair of text date strings, e.g. c('1981-02-05:12','1981-03-27:18') - inclusive.}

\item{type}{- 'mean' (default), 'spread', 'normal', or 'standard.deviation'.
Note that standard deviations are not available over opendap.}

\item{height.range}{Bottom and top heights in hPa - leave NULL for monolevel}

\item{lat.range}{Min and max latitude in degrees - defaults to c(-90,90)}

\item{lon.range}{Min and max longitude in degrees - defaults to c(-180,180)}

\item{opendap}{must be TRUE - no local option}
}
\value{
A GSDF field with the selected multidimensional data
}
\description{
Up to 4d (lat, lon, height, time).
}
\details{
MERRA data is in daily files - time range must be contained in 1 day
}

