\name{MERRA.show.variables}
\alias{MERRA.show.variables}
\title{Merra show variables}
\usage{
  MERRA.show.variables()
}
\description{
  List all the variables available in the MERRA reanalyis.
}
\details{
  Two sorts of variables: \describe{ \item{Monolevel} {Lat
  x Lon x Time} \item{Pressure level} {Lat x Lon x Height x
  Time} } This function has no parmeters and returns no
  value it prints a list of available variables.
}

